+++ app/Http/Controllers/Admin/UsersController.php
@@ -11,6 +11,8 @@ use App\Models\User;
 use Gate;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Alert;
+use Auth;
 
 class UsersController extends Controller
 {
@@ -37,6 +39,7 @@ class UsersController extends Controller
         $user = User::create($request->all());
         $user->roles()->sync($request->input('roles', []));
 
+        Alert::success('تم بنجاح', 'تم إضافة المستخدم بنجاح ');
         return redirect()->route('admin.users.index');
     }
 
@@ -56,6 +59,7 @@ class UsersController extends Controller
         $user->update($request->all());
         $user->roles()->sync($request->input('roles', []));
 
+        Alert::success('تم بنجاح', 'تم تعديل بيانات المستخدم بنجاح ');
         return redirect()->route('admin.users.index');
     }
 
@@ -72,9 +76,16 @@ class UsersController extends Controller
     {
         abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
+
+        if($user->id == Auth::id()){ 
+            Alert::error('لايمكن حذف هذا المستخدم');
+            return 0;
+        }
+
         $user->delete();
 
-        return back();
+        Alert::success('تم بنجاح', 'تم  حذف المستخدم بنجاح ');
+        return 1;
     }
 
     public function massDestroy(MassDestroyUserRequest $request)
